backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
MarkupSafe==3.0.3
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.2.1
rave-python==1.4.2
redis==7.0.0
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
